助教的計時方式
