+++ README.md
@@ -11,6 +11,7 @@
   |----adj_matrix.cpp
   |----adj_matrix_pointer.cpp
   |----adj_list_pointer.cpp
+  |----main_time_calculating.cpp
   |----test_cases
      |----11.txt
      |----12.txt
@@ -36,17 +37,41 @@
      |----open_benchmarks_1.jpg
      |----open_benchmarks_2.jpg
   |----test_cases
-     |----adj_list.cpp
-     |----adj_matrix.cpp
-     |----adj_matrix_pointer.cpp
-     |----adj_list_pointer.cpp
+     |----adj_list_time.cpp
+     |----adj_matrix_time.cpp
+     |----adj_matrix_pointer_time.cpp
+     |----adj_list_pointer_time.cpp
 ```
 ## input file:
 ```
 /test_cases/test_adj_matrix.txt
+for adj_matrix.cpp
+    adj_matrix_pointer.cpp
+    adj_matrix_time.cpp
+    adj_matrix_pointer_time.cpp
+
 /test_cases/test_adj_list.txt
+for adj_list.cpp
+    adj_list_time.cpp
+
+/test_cases/test_adj_list_for_pointer.txt
+for adj_list_pointer.cpp
+    adj_list_pointer_time.cpp
+
 /open_benchmarks/test_adj_matrix.txt
+for adj_matrix.cpp
+    adj_matrix_pointer.cpp
+    adj_matrix_test.cpp
+    adj_matrix_pointer_test.cpp
+    
 /open_benchmarks/test_adj_list.txt
+for adj_list.cpp
+    adj_list_time.cpp
+
+/open_benchmarks/test_adj_list_for_pointer.txt
+for adj_list_pointer.cpp
+    adj_list_pointer_time.cpp
+
 ...
 ```
