+++ adj_matrix_pointer.cpp
@@ -84,8 +84,6 @@ int Load_graph(int ***&adjacency_lists, int *&number_of_vertexes)
 
 int main()
 {
-	clock_t t1, t2;
-
 	int number_of_graphs;
 	int ***data;
 	int *number_of_vertexes;
@@ -93,13 +91,10 @@ int main()
 
 	for (int i = 0; i < number_of_graphs; i++)
 	{
-		t1 = clock();
 		if (is_Bipartite(i, number_of_vertexes[i], data))
 			cout << "True\n";
 		else
 			cout << "False\n";
-		t2 = clock();
-		printf("%lfseconds\n", (t2 - t1) / (double)(CLOCKS_PER_SEC));
 	}
 
 	return 0;
