+++ test_time/adj_list_pointer_time.cpp
@@ -82,28 +82,34 @@ int Load_graph(int ***&adjacency_lists, int *&number_of_vertexes)
 	return 0;
 }
 
-struct timespec diff(struct timespec start, struct timespec end) {
-  struct timespec temp;
-  if ((end.tv_nsec-start.tv_nsec)<0) {
-    temp.tv_sec = end.tv_sec-start.tv_sec-1;
-    temp.tv_nsec = 1000000000+end.tv_nsec-start.tv_nsec;
-  } else {
-    temp.tv_sec = end.tv_sec-start.tv_sec;
-    temp.tv_nsec = end.tv_nsec-start.tv_nsec;
-  }
-  return temp;
+// for calculating the execution time of this program
+struct timespec diff(struct timespec start, struct timespec end)
+{
+	struct timespec temp;
+	if ((end.tv_nsec - start.tv_nsec) < 0)
+	{
+		temp.tv_sec = end.tv_sec - start.tv_sec - 1;
+		temp.tv_nsec = 1000000000 + end.tv_nsec - start.tv_nsec;
+	}
+	else
+	{
+		temp.tv_sec = end.tv_sec - start.tv_sec;
+		temp.tv_nsec = end.tv_nsec - start.tv_nsec;
+	}
+	return temp;
 }
 
 int main()
 {
-	// 儲存時間用的變數
-	struct timespec start, end;
-	double time_used;
+	///////////////////////////////////////////////////////////////////////////////
+	// to calculate the total execution time of one case
+	struct timespec time_start, time_end;
+	double time_used_total;
+	clock_gettime(CLOCK_MONOTONIC, &time_start);
 
-	// 計算開始時間
-	clock_gettime(CLOCK_MONOTONIC, &start);
-
-	//clock_t t1, t2;
+	struct timespec time_sub_1, time_sub_2;
+	double time_used_sub;
+	///////////////////////////////////////////////////////////////////////////////
 
 	int number_of_graphs;
 	int ***data;
@@ -111,22 +117,34 @@ int main()
 	number_of_graphs = Load_graph(data, number_of_vertexes);
 	for (int i = 0; i < number_of_graphs; i++)
 	{
-		//t1 = clock();
+		///////////////////////////////////////////////////////////////////////////////
+		// to calculate the execution time
+		clock_gettime(CLOCK_MONOTONIC, &time_sub_1);
+		///////////////////////////////////////////////////////////////////////////////
+
 		if (is_Bipartite(i, number_of_vertexes[i], data))
 			cout << "True\n";
 		else
 			cout << "False\n";
-		//t2 = clock();
-		//printf("%lfseconds\n", (t2 - t1) / (double)(CLOCKS_PER_SEC));
+
+		///////////////////////////////////////////////////////////////////////////////
+		// to calculate the execution time of graph
+		clock_gettime(CLOCK_MONOTONIC, &time_sub_2);
+		struct timespec temp = diff(time_sub_1, time_sub_2);
+		time_used_sub = temp.tv_sec + (double)temp.tv_nsec / 1000000.0;
+		cout << "The program execution time of this case is: " << time_used_sub << "ms" << endl;
+		cout << "-----------------------------------------------------" << endl;
+		///////////////////////////////////////////////////////////////////////////////
 	}
 
-	// 計算結束時間
-	clock_gettime(CLOCK_MONOTONIC, &end);
-
-	// 計算實際花費時間
-	struct timespec temp = diff(start, end);
-	time_used = temp.tv_sec + (double) temp.tv_nsec / 1000000000.0;
-	printf("Time = %f\n", time_used);
+	///////////////////////////////////////////////////////////////////////////////
+	// to calculate the total execution time
+	cout << endl;
+	clock_gettime(CLOCK_MONOTONIC, &time_end);
+	struct timespec temp = diff(time_start, time_end);
+	time_used_total = temp.tv_sec + (double)temp.tv_nsec / 1000000.0;
+	cout << "The program execution time of this program is: " << time_used_total << "ms" << endl;
+	///////////////////////////////////////////////////////////////////////////////
 
 	return 0;
 }
